iox-#1196 Fix useless cast warnings

+++ iceoryx_hoofs/include/iceoryx_hoofs/internal/posix_wrapper/message_queue.hpp
@@ -49,9 +49,9 @@ class MessageQueue : public DesignPattern::Creation<MessageQueue, IpcChannelErro
   public:
     static constexpr mqd_t INVALID_DESCRIPTOR = std::numeric_limits<mqd_t>::max();
     static constexpr int32_t ERROR_CODE = -1;
-    static constexpr size_t SHORTEST_VALID_QUEUE_NAME = 2;
-    static constexpr size_t NULL_TERMINATOR_SIZE = 1;
-    static constexpr size_t MAX_MESSAGE_SIZE = 4096;
+    static constexpr uint64_t SHORTEST_VALID_QUEUE_NAME = 2;
+    static constexpr uint64_t NULL_TERMINATOR_SIZE = 1;
+    static constexpr uint64_t MAX_MESSAGE_SIZE = 4096;
 
     /// for calling private constructor in create method
     friend class DesignPattern::Creation<MessageQueue, IpcChannelError>;
